requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --allow-unsafe --no-annotate
#
aws-requests-auth==0.4.3
beautifulsoup4==4.12.2
boto3==1.26.138
botocore==1.29.138
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-spinner==0.1.10
configparser==5.3.0
cryptography==40.0.2
idna==3.4
importlib-metadata==6.6.0
jaraco-classes==3.2.3
jeepney==0.8.0
jmespath==1.0.1
keyring==23.13.1
lxml==4.9.2
more-itertools==9.1.0
polling2==0.5.0
pycparser==2.21
pysocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
ruamel-yaml==0.17.26
ruamel-yaml-clib==0.2.7
s3transfer==0.6.1
secretstorage==3.3.3
semver==3.0.0
six==1.16.0
soupsieve==2.4.1
tabulate==0.9.0
urllib3==1.26.15
zipp==3.15.0
